[cocoa] Restructure stubs, split by component
