+++ renderer_cocoa/lib/stubs/brisk_cocoa_view.h
@@ -1,0 +1,11 @@
+#ifndef BRISK_COCOA_VIEW_H
+#define BRISK_COCOA_VIEW_H
+
+#define CAML_NAME_SPACE
+
+#include "brisk_cocoa.h"
+
+#define Val_View(v) ((value)(v))
+#define View_val(v) ((__bridge View *)(value)(v))
+
+#endif /* BRISK_COCOA_VIEW_H */
